state: add Life to state.Environment
Environment has been given new methods
to mark an environment as dying, and
to remove the environment from state.
An environment must now exist in state
and be Alive for services or machines
to be added to state.

A secondary change here is to use the
environment's UUID in the tag, rather
than the environment's name, which is
non-unique.