names: NewActionTag must panic if the action tag is not valid
The contract of the NewXXXTag methods is they should return a valid names.Tag impl or they should panic. As these functions are called almost exclusively in tests to create fixtures, this black and white pattern is appropriate.

This PR changes names.NewActionTag to the expected behavior and adds tests to assert this same.