s/Disk/Volume/, optional machine scope
DiskTag is renamed to VolumeTag, to reflect
the entity name we use in the juju/juju code.
The tags are not yet used in a released version,
so we can do this without breakage.

This change also introduces an optional
machine ID prefix to volume IDs, identifying
the machine that a volume is scoped to, if any.