Granular validation errors
The names package is a great way to check if a name is valid, but really
struggles to tell you why a given name is valid. The following change
starts to do that by exposing a ValidateApplicationName, which will tell
you why a given name is wrong.

The code is rather dull, but is effective in that we try and be helpful
upfront (uppercase issues or numbers after the last hyphen), then just
point out an issue in a more programtic sense. Whilst the last is great
if you're a programmer, it's more cryptic for general users.