chore: add name, source, revision and arch to charm tag
This patch refactors and updates the charm tag.
First, it updates the validation regular expressions to match our
current acceptance of the charm URLs and removes legacy accepted ones
(like cs: and ~user).
Second, it adds methods (and inner fields) to return the charm's Name,
Source, Architecture (optional) and Revision.