feat: support UUID v7 tags for spaces and subnets
A bug has been discovered recently in Juju 4.0 where we cannot move
spaces because the subnet (which is automatically retrieved and inserted
into the db with a new UUID v7 as primary key) cannot be passed as a new
Subnet tag and therefore it panics.

The fix is to add a new validation regex for spaces and subnets which
validate UUIDs v7 and keep the legacy (and deprecated) sequence ID as
fallback validation for backward compatibility.