fix: space tag shouldn't validate against UUIDs
In a previous patch, the validation against UUIDs v7 was added for
subnets and spaces. It turns out it wasn't needed for spaces since we
use the space *name* as a tag in Juju.
This wrong validation was removed and the test was updated to reflect
that we use a space name instead of a sequence number.

Bonus: the fallback validation for sequence numbers was removed from
subnets, meaning that now only UUIDs v7 are valid. Nothing should break
from this change since Juju already only uses UUIDs v7 for subnet IDs.