fix: add fallback for spaces and subnets validation
This patch adds back a fallback for validating legacy space and subnet
formats.
This fallback was added first in https://github.com/juju/names/pull/118
then incorrectly removed in https://github.com/juju/names/pull/119 and
now it's added back again. The reason is that if a 4+ Juju client tries
to `move-to-space` on a 3.x controller, the client will panic. This
fallback prevents that.